backend/requirements.txt
python-socketio
flask
flask-socketio
eventlet
